size_t + iostream

+++ gladyshev.matvey/P1/counterreversal.cpp
@@ -24,7 +24,7 @@ void gladyshev::CounterReversal::operator()(long long int number_)
   }
   sup_number_ = number_;
 }
-unsigned long long int gladyshev::CounterReversal::operator()() const
+size_t gladyshev::CounterReversal::operator()() const
 {
   return counter_sign_;
 }
