+++ gladyshev.matvey/P1/counterreversal.h
@@ -1,15 +1,17 @@
 #ifndef COUNTERREVERSAL_H
 #define COUNTERREVERSAL_H
 
+#include <iostream>
+
 namespace gladyshev
 {
   struct CounterReversal
   {
     CounterReversal();
     void operator()(long long int number_);
-    unsigned long long int operator()() const;
+    size_t operator()() const;
   public:
-    unsigned long long counter_sign_;
+    size_t counter_sign_;
   private:
     long long int number_;
     long long int sup_number_;
